Adds `token` option to `Joken.Plug` (#167)
`token`, like `verify`, can be set to a functon that should return a compact JWT token.  This allows for flexibility in where the token is coming from if not in the `Bearer` header.

The default action (locating the token in the `Bearer` header is preserved if no token function is provided.