Enhance token config and add docs
This switches from passing the token config in an option to a
function. Joken.Config is now a behaviour with overridable callbacks
so that the implementation is cleaner.

Also removed the core implementation from the using macro to
Joken.Config. This follows best practices of less work on macros.

Enhanced the documentation of modules overall but haven't touched
README yet.