Joken.Config free API
This commit has the introduction of a Joken.Config free API usage.
It is easier to migrate old style to this style if needed. You can
build the token configuration by hand and pass it to Joken functions.

This also introduces the multiple hooks on Config discussed before.

It cleans Joken.Error and the test database to raise de coverage to
95%.