Access current_time_adapter in runtime instead of compile time
This fixes issue #236
I tried using Mox to inject a Joken.CurrentTime.Mock instance but because the way Joken was accessing the adapter using module attributes which are resolved at compile time instead of runtime. And because the Joken module is already compiled when we pull it as a dependency, the adapter was always using the `Joken.CurrentTime.OS` in test.

By using a private function to access the adapter, we make sure that the adapter is resolved at runtime.